Consolidate Python scripts into unified CLI tool
- Created cli.py with 4 subcommands: scrape-pubs, update-cv, generate-pdf, enrich-inventory
- Removed individual script files (semantic_scholar_scraper.py, update_cv.py, generate_cv_pdf.py, enrich_inventory.py)
- Added _scripts/README.md with complete CLI documentation
- Updated GitHub Actions workflow to use new CLI commands
- Updated main README with new script structure and usage examples
- All functionality preserved with cleaner interface
